+++ src/Behat/MinkExtension/ServiceContainer/MinkExtension.php
@@ -37,10 +37,8 @@ use Symfony\Component\DependencyInjection\Reference;
  *
  * @author Konstantin Kudryashov <ever.zet@gmail.com>
  * @author Christophe Coevoet <stof@notk.org>
- *
- * @final since 2.8.0
  */
-class MinkExtension implements ExtensionInterface
+final class MinkExtension implements ExtensionInterface
 {
     const MINK_ID = 'mink';
     const SELECTORS_HANDLER_ID = 'mink.selectors_handler';